Major fix with large databases: READ DESC
Databases with 10k+ pages start to slow down the entire site by writing it to the db.txt file, so once it gets large enough, the db.txt file is written to every 30 seconds and it takes 30 seconds to write, so the entire site doesn't work. This patch uppers that so it only hiccups every once in a "while"

I will be working on a better approach that auto-scales and makes the server handle it better.